Change to replace `Buffer` with `Uint8Array`
Previously, this project depended on Node’s buffer.
This commit changes that: it replaces that support with `Uint8Array`.
In most cases, this will be fine, because the Node `Buffer` class
subclasses `Uint8Array`.
However, there are differences in which encodings are supported
when turning this binary data into a string, and there are differences in
how methods work.